+++ robot_sf/gym_env/reward.py
@@ -41,7 +41,7 @@ def simple_reward(
     if meta["is_robot_at_goal"]:
         reward += reach_waypoint_reward
 
-    return reward
+    return float(reward)
 
 
 def simple_ped_reward(
@@ -90,7 +90,7 @@ def simple_ped_reward(
     if meta["is_robot_at_goal"]:
         reward += robot_at_goal_penalty
 
-    return reward
+    return float(reward)
 
 
 def punish_action_reward(
@@ -129,4 +129,4 @@ def punish_action_reward(
         if action_diff > 0:
             reward += punish_action_penalty * action_diff
 
-    return reward
+    return float(reward)